pulling code out of aws_stuff_doer.py, using it for main function & creating libraries
