+++ src/aws_stuff_doer/open/__init__.py
@@ -1,0 +1,1 @@
+"""Open library for AWS SSO."""
