+++ src/aws_stuff_doer/uilib/screens/help_screen/HelpScreen.py
@@ -1,8 +1,9 @@
 from textual.app import ComposeResult
 from textual.screen import Screen
 from textual.widgets import Header, Footer, Static
+from textual.command import CommandList
 
-HELP_TEXT =  """
+HELP_TEXT = """
                            Help Menu
     ASD: An AWS Utility to help manage your projects and sso sessions.
 
@@ -14,11 +15,11 @@ HELP_TEXT =  """
 
     """
 
-class HelpScreen(Screen): # type: ignore
+
+class HelpScreen(Screen):  # type: ignore
     """A simple help screen."""
 
     def compose(self) -> ComposeResult:
         yield Header()
         yield Static(HELP_TEXT)
         yield Footer()
-
