+++ src/aws_stuff_doer/uilib/widgets/__init__.py
@@ -1,0 +1,3 @@
+"""The __init__.py filr for the widgets package."""
+
+from .profile_select import ProfileSelect  # type: ignore
